Check for non-existent authorities before modification and fix user deletion candidate collection

+++ repository/src/main/java/de/acosix/alfresco/mtsupport/repo/sync/GroupCreationAndParentRemovalWorker.java
@@ -75,8 +75,12 @@ public class GroupCreationAndParentRemovalWorker extends AbstractSyncBatchWorker
             {
                 for (final String parent : parents)
                 {
-                    LOGGER.debug("Removing {} from group {}", groupShortName, parent);
-                    this.authorityService.removeAuthority(parent, group);
+                    if (this.authorityService.authorityExists(parent))
+                    {
+                        LOGGER.debug("Removing {} from group {}", groupShortName, parent);
+                        this.authorityService.removeAuthority(parent, group);
+                    }
+                    // else: not a problem
                 }
             }
         }
